BUG: When unregistering a data file from the data update service, a collection modified exception was thrown.
By using Linq's `Where(predicate)` method, the Enumerable was intrinsicly linked to the private List. This meant that the result could not be iterated over when to remove an element from the private List. Instead, the result was unlinked from the private List by converting to a list.
A unit test has also been added to check this bug.