Merged PR 2087: BUG: Fixed a few more bugs with auto update.
- BUG: DataUpdateService was using file creation date as an indicator of the published date of the data in the file when an exact date was not provided.
When an updated file was copied over the old one, creation date was not updated, leading to uneccessary downloads of the same data in the future.
The creation time is now set when the new file is copied over the old one.
Additionally, some update methods are based on the user copying a new file over the old one. The DataUpdateService then detects the change and refreshes the engine. In these cases, we cannot rely on file creation date so use last write time instead.
- BUG: If UpdateOnStartup is enabled and no update was needed or update failed then a timer would not be created to check for updates in future. This has now been corrected.
- TEST: Update DataUpdateService tests to use last write date rather than creation date.
- TEST: Also added a test to verify the update on startup but no update needed scenario described above.