BUG: Cloud engines will now use reflection to determine property type rather than relying on the data from the cloud response.
BUG: Cloud engines will now use reflection to determine property type rather than relying on the data from the cloud response.
This is needed due to cloud responses potentially containing collections with many different types of items.

If a property cannot be found using reflection then a warning will be logged and the meta-data for that property will not be available.
If no properties can be found for a given 'product' then an exception will be thrown.