REFACTOR: JsonBuilderElement now has separate methods for GetValues and GetJsonValuesForProperty.
Split the GetValues method so that a separate method 'AddJsonValuesForProperty' does the work of populating the json property dictionary based on the property value and the delay execution config.
This allows the cloud implementation to reuse the majority of the functionality while still being able to hook into the parts it needs.