BUG: Fix several bugs with the ASP.NET integration and example.
All these changes relate to the ASP.NET web integration.

BUG: Correct a [bug](https://github.com/51Degrees/pipeline-dotnet/issues/4) that causes a crash when using the ASP.NET integration
BUG: ClientSideEvidenceEnabled setting should default to true.
BUG: The Sequence, JsonBuilder and JavaScript builder elements are not always added in the correct locations in the pipeline.
REFACTOR: The majority of the code in the FiftyOneJsProvider class is duplicating what is in ClientsidePropertyService for the ASP.NET Core integration. This duplicated code is moved to the Web.Shared library.
TEST: Added automated tests for the new shared ClientsidePropertyService.
FEAT: The ASP.NET integration had not been updated to use the ‘callback’ mechanism by default. This was previously implemented for the ASP.NET Core integration in 4.1.10.
EXAMPLE: The configuration for the ASP.NET example was incorrect.
EXAMPLE: The ASP.NET example needs to load the relevant assemblies into the app context in order to allow the pipeline to build the required flow elements.
EXAMPLE: Added illustrative configuration to the web integration examples in order to show the various settings available.