FEAT: Add settings to allow all evidence values to be shared by usage sharing element.
FEAT: Add settings to allow all evidence values to be shared by usage sharing element.
- The ShareUsageElement.SetShareAllEvidence method can be used to specify that all evidence should be shared. This short circuits the usual checks and just shares everything.
- The ShareUsageElement.SetShareAllQueryStringParameters method can be used to specify that all query string parameters should be shared, rather than having to maintain a whitelist.

TEST: Added tests to verify this new functionality.
FEAT: The HostAddress property on ShareUsageElementBase is now protected instead of private. This allows it to be reused by the cloud sharing class, avoiding the need to re-implement it. (Resolves a TODO in that class)