FEAT: Add parameters from POST body to the evidence and add evidence to the cloud request using a defined order of precedence.
The change is to implement evidence key precedence for the cloud request engine .
The precedence is:
1. Query
2. Header
3. Cookie
4. The rests in in alphabetical order.

Query evidence override all other evidence keys but if there is a conflict between non-query evidence then the evidence with the highest order of precedence will be selected and a warning logged. An example of conflict might be 'header.User-Agent' and 'cookie.User-Agent'.