FEAT: CloudAspectEngineBase implements process method which handles retrieval of json response from CloudRequestEngine.
The CloudAspectEngineBase class now implements ProcessEngine. This implements common code for all CloudAspectEngines which checks for the existence of a CloudRequestEngine, checks that it has run before the CloudAspectEngine and also checks if the CloudRequestEngine has populated the JsonResponse.

If all the checks are ok then the CloudAspectEngineBase class call the ProcessCloudEngine method which should be implemented by dervied classes.

CloudAspectEngines should now implement the abstract ProcessCloudEngine method which is only called if the cloud JsonResponse is set.