Merged PR 7917: BUG: Better handling for missing properties during calls to cloud service
OPTIM: Log a warning in the set headers element if the property is missing, instead of throwing an exception.

This can be the case if a cloud API call has failed. The device data will not have been processed, so a PropertyMissingException will be thrown when trying to access one.
This is now handled more cleanly.
#5237

Related work items: #5237