Merged Pull Request '#174 version/feature/get-flow-elements->version/4.5: Change FlowElements to return ALL flow elements, including sub elements'
* BUG: Change FlowElements to return ALL flow elements. Before, subelements that are part of a parallel element would not be returned as the Parallel element would not be unpacked and will not cobntain an element data key. Now, unpack the parallel elements to create the full list.
Add extended test methods for the GetWhere method to ensure that all element data for all elements (incfluding parallel elements is returned). These have been added to a seperate file/test class as the set up for these tests differs from the setup for the other flowdata tests.

* CLEANUP 80 Chars

* FEAT/CLEANUP: Recursively add sub elements to the list.
Add new test to check that nest parallel elements are returned.
Update documentation.

* FEAT: Instead of Recursion, use a Queue to avoid possible stack overflow. Aslo add a check to make sure the element.isPublic before adding it to the list.
Change the Tests to make sure the elements are not checkd in an ordered fashion.